Bugfix: Use triple equals, not double equals
Using double equals meant that the list of array keys was testing
equal even though it wasn't. I think this was because it had a single
item which wasn't numeric, so it would become `0` when cast to a
number. As a result of the mistaken conversion, the indexing on the
next line was failing, because that key did not actually exist.